+++ src/ElfParser.c
@@ -1,5 +1,5 @@
 #include "ElfParser.h"
-#include "ElfReader.h"
+#include "ElfIO.h"
 #include "ElfStringTable.h"
 #include <stddef.h>
 #include <stdlib.h>
