+++ test/testElfReader.c
@@ -1,4 +1,4 @@
-#include "ElfReader.h"
+#include "ElfIO.h"
 #include "UnitTest.h"
 #include "elf.h"
 #include <stdlib.h>
